add various type hints, use more envs in training

+++ robot_sf/extenders_py_sf/simulation_config.py
@@ -25,7 +25,7 @@ class SimulatorConfigFlags:
 @dataclass
 class ActionPool:
     actions: List[str]
-    probs_in_percent: List[int]
+    probs_in_percent: List[float]
 
 
 @dataclass
